Added decoder error cases for Playback Error
These cases could happen when the hardware decoder of the device is already occupied, for example when you tried to run multiple videos at once or when a video player is not properly deallocated. This was common in any iOS devices such as iPhone because it has limited hardware decoders available while there was an memory leak that is caused by a nasty retain cycle (which is already fixed in the recent PR).

Perhaps it should be better to add some inline docs to these error cases to let developers know what these error cases means, but I dare not to do so yet.